+++ RELEASE-NOTES.md
@@ -4,6 +4,7 @@
 Unrealed. Updated by Steviegt6
 
 * Compiles directly to x86 instead of AnyCPU due to the actual StardewValley dependency being a strictly-32bit program.
+* No longer depends on SMAPI, still retains SMAPI patches.
 
 ## 1.1.7
 Released 8 May 2021. Updated by Steviegt6.