+++ RELEASE-NOTES.md
@@ -5,6 +5,7 @@ Unrealed. Updated by Steviegt6
 
 * Compiles directly to x86 instead of AnyCPU due to the actual StardewValley dependency being a strictly-32bit program.
 * No longer depends on SMAPI, still retains SMAPI patches.
+* Added ability to depot-download using the program + install with that download without having to exit or enter a path.
 
 ## 1.1.7
 Released 8 May 2021. Updated by Steviegt6.