+++ RELEASE-NOTES.md
@@ -1,7 +1,7 @@
 ﻿← [README](README.md)
 
 ## 1.1.8
-Unrealed. Updated by Steviegt6
+Released 2 July 2021. Updated by Steviegt6
 
 * Compiles directly to x86 instead of AnyCPU due to the actual StardewValley dependency being a strictly-32bit program.
 * No longer depends on SMAPI, still retains SMAPI patches.